+++ Gruntfile.js
@@ -1,53 +1,44 @@
 module.exports = function(grunt) {
 	'use strict';
 
-	require('matchdep').filterDev('grunt-*').forEach(grunt.loadNpmTasks);
+	require('load-grunt-tasks')(grunt);
 
 	grunt.initConfig({
 
 		watch: {
 
-			styles: {
-				files: ['assets/scss/**/*.{scss,sass}'],
-				tasks: ['styles'],
-				options: {
-					debounceDelay: 500
-				}
+			css: {
+				files: ['assets/scss/**/*.scss'],
+				tasks: ['css']
 			},
 
-			scripts: {
+			js: {
 				files: ['assets/js/**/*.js'],
-				tasks: ['scripts'],
-				options: {
-					debounceDelay: 500
-				}
+				tasks: ['js']
 			}
 
 		},
 
 		jshint: {
+			options: {
+				reporter: require('jshint-stylish')
+			},
 			gruntfile: ['Gruntfile.js'],
-			assets: ['assets/js/**/*.dev.js']
+			assets: ['assets/js/**/*.js']
 		},
 
-		uglify: {
-			dist: {
-				files: {
-					'assets/js/admin-single.min.js': ['assets/js/admin-single.js']
-				}
-			}
-		},
-
-		compass: {
-			dist: {
-				options: {
-					config: 'assets/config.rb'
-				}
+		sass: {
+			assets:  {
+				cwd: 'assets/scss',
+				src: '*.scss',
+				dest: 'assets/css',
+				expand: true,
+				ext: '.css'
 			}
 		},
 
 		autoprefixer: {
-			dist: {
+			assets: {
 				expand: true,
 				flatten: true,
 				src: 'assets/css/*.css',
@@ -56,7 +47,7 @@ module.exports = function(grunt) {
 		},
 
 		csso: {
-			dist: {
+			assets: {
 				expand: true,
 				flatten: true,
 				src: 'assets/css/*.css',
@@ -65,35 +56,21 @@ module.exports = function(grunt) {
 		},
 
 		imagemin: {
-			dist: {
-				options: {
-					optimizationLevel: 7,
-					progressive: true
-				},
+			assets: {
 				files: [{
 					expand: true,
 					cwd: 'assets/images/',
 					src: '**/*',
 					dest: 'assets/images/'
 				}]
-			}
-		},
-
-		includes: {
-			options: {
-				includePath: 'vendor',
-				includeRegexp: /\/\/\s*import\s+['"]?([^'"]+)['"]?\s*?$/,
 			},
-			scripts: {
-				src: ['assets/js/admin-single.dev.js'],
-				dest: 'assets/js/admin-single.js'
-			},
-			styles: {
-				src: ['assets/css/admin-single.css'],
-				dest: 'assets/css/admin-single.css',
-				options: {
-					includeRegexp: /\/\*\s*import\s+['"]?([^'"]+)['"]?\s*\*\/?$/
-				}
+			screenshots: {
+				files: [{
+					expand: true,
+					cwd: 'screenshots/',
+					src: '**/*',
+					dest: 'screenshots/'
+				}]
 			}
 		},
 
@@ -113,7 +90,16 @@ module.exports = function(grunt) {
 						'includes/**/*',
 						'admin/**/*',
 						'assets/**/*',
-						'languages/**/*'
+						'languages/**/*',
+
+						// CodeMirror
+						'vendor/codemirror/lib/codemirror.css',
+						'vendor/codemirror/lib/codemirror.js',
+						'vendor/codemirror/mode/clike/clike.js',
+						'vendor/codemirror/mode/php/php.js',
+						'vendor/codemirror/addon/search/searchcursor.js',
+						'vendor/codemirror/addon/search/search.js',
+						'vendor/codemirror/addon/edit/matchbrackets.js'
 					],
 					dest: 'deploy/plugin',
 					filter: 'isFile'
@@ -133,6 +119,16 @@ module.exports = function(grunt) {
 			}
 		},
 
+		wp_deploy: {
+			release: {
+				options: {
+					plugin_slug: 'code-snippets',
+					svn_user: 'bungeshea',
+					build_dir: 'deploy/plugin'
+				},
+			}
+		},
+
 		shell: {
 
 			/**
@@ -145,11 +141,12 @@ module.exports = function(grunt) {
 
 	});
 
-	grunt.registerTask( 'styles', ['compass', 'includes:styles', 'autoprefixer', 'csso'] );
-	grunt.registerTask( 'scripts', ['jshint', 'includes:scripts'] );
+	grunt.registerTask( 'css', ['sass', 'autoprefixer', 'csso'] );
+	grunt.registerTask( 'js', ['jshint'] );
 
 	grunt.registerTask( 'deploy', ['clean:deploy', 'copy:plugin', 'copy:assets'] );
 	grunt.registerTask( 'phpdoc', 'shell:phpdoc' );
 
-	grunt.registerTask( 'default', ['styles', 'scripts', 'uglify'] );
+	grunt.registerTask( 'default', ['css', 'js'] );
+	grunt.registerTask( 'release', ['default', 'deploy', 'wp_deploy'] );
 };